+++ test/cannot-find-exception.js
@@ -1,7 +1,7 @@
-const fs = require('fs')
-const path = require('path')
-const tap = require('tap')
-const clearFolder = require('../cli.js')
+import fs from 'fs'
+import path from 'path'
+import tap from 'tap'
+import cf from '../cli.js'
 
 const folderErrObj = function ( folder ) {
     return {
@@ -16,7 +16,7 @@ const skipObj = {
 
 tap.test(`a message is logged when a non-existent folder is specified`, function ( t ) {
     const folder = 'non-existent-folder-name'
-    const throwing = function () { clearFolder( [folder] ) }
+    const throwing = function () { cf( [folder] ) }
     t.throws( throwing, folderErrObj( folder ), skipObj )
 
     t.end()
@@ -24,7 +24,7 @@ tap.test(`a message is logged when a non-existent folder is specified`, function
 
 tap.test(`a message is logged when the home-folder is referenced`, function ( t ) {
     const folder = '~/.ssh'
-    const throwing = function () { clearFolder( [folder] ) }
+    const throwing = function () { cf( [folder] ) }
     t.throws( throwing, folderErrObj( folder ), skipObj )
 
     t.end()